added delet functions and alerts

+++ FRONTEND/src/pages/admin/Users.jsx
@@ -1,6 +1,7 @@
 import React, { useState, useEffect } from 'react'; // Import the 'useState' and 'useEffect' hooks from React.
 import axios from 'axios'; // Import the 'axios' library for making HTTP requests.
 import getAllUsers from '../../components/GetAllUsers'; // Import the 'getAllUsers' function from the 'GetAllUsers' module.
+import toast, { Toaster } from 'react-hot-toast';
 
 const Users = () => {
   const [users, setUsers] = useState([]); // Create a state variable 'users' and a setter function 'setUsers' with an initial value of an empty array.
@@ -11,6 +12,7 @@ const Users = () => {
       .catch((error) => console.error('Error fetching users:', error)); // Log an error message if an error occurs during the request.
   }, []); // The empty array passed as the second argument to 'useEffect' ensures that the side effect runs only once after the initial render.
 
+  // UPDATE USER-------------------------------------
   const updateUser = async (user) => { // Define an 'updateUser' function that takes a 'user' object as an argument.
     try {
       const response = await axios.put(`http://localhost:5000/api/users/${user._id}`, user); // Send a PUT request to update the user data.
@@ -23,11 +25,18 @@ const Users = () => {
     // Define a 'handleSave' function that takes a 'user' object as an argument.
   const handleSave = (user) => {
     updateUser(user)
-      .then(() => getAllUsers())
+      .then(() => {
+        toast.success('User updated successfully');// Display a success message using the 'toast' library.
+        setTimeout(() => {
+          window.location.reload();
+        }, 2000);// Reload the page after 2 seconds
+      })
+      .then(() => getAllUsers())// Call the 'getAllUsers' function to fetch the updated user data.
       .then((data) => setUsers(data))
       .catch((error) => console.error('Error saving user:', error));
   };
 
+  // DELETE USER-------------------------------------
   // Define a 'deleteUser' function that takes a user 'id' as an argument.
   const deleteUser = async (id) => {
     try {
@@ -39,16 +48,25 @@ const Users = () => {
   };
 
   const handleDelete = (id) => {
-    deleteUser(id)
-      .then(() => getAllUsers())
-      .then((data) => setUsers(data))
-      .catch((error) => console.error('Error deleting user:', error));
+    if (window.confirm('Are you sure you want to delete this user?')) {
+      deleteUser(id)
+        .then(() => {
+          getAllUsers()
+            .then((data) => setUsers(data));
+          toast.success('User deleted');
+        })
+        .catch((error) => {
+          console.error('Error deleting user:', error);
+          toast.error('Error deleting user');
+        });
+    }
   };
 
 
 
   return (
     <div className="text-gray-900 bg-gray-200">
+      <Toaster />
       <div className="p-4 flex">
         <h1 className="text-3xl">Users</h1>
       </div>
